migrate some tests away from AssetGroup (#12606)
### Summary & Motivation

`AssetGroup` is no longer part of the public API, but we use it under
the covers. I'd like to remove it. A lot of the tests on it are still
useful though. So I'm moving them away from `AssetGroup`.

This is a random selection of tests that I was able to move in a bounded
amount of time. The goal is to eventually delete test_asset_group.py and
then delete asset_group.py.

### How I Tested These Changes
